requirements.txt
# for pinecone

aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepdiff==7.0.1
Django==5.1
dnspython==2.6.1
emoji==2.12.1
et-xmlfile==1.1.0
fastapi==0.114.2
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.0
langchain-cli==0.0.30
langchain-community==0.3.0
langchain-core==0.3.0
langchain-mistralai==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langserve==0.3.0
langsmith==0.1.120
libcst==1.4.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
mistralai==0.4.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.19.0
opencv-python==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pinecone-client==2.2.4
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==5.27.3
psutil==6.0.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==4.3.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.1
sse-starlette==1.8.2
starlette==0.38.5
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.20.0
tomlkit==0.12.5
tqdm==4.66.5
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.5
unstructured-client==0.25.5
urllib3==2.2.2
uvicorn==0.23.2
win32-setctime==1.1.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4




# # for faiss
# aiohappyeyeballs==2.3.7
# aiohttp==3.10.4
# aiosignal==1.3.1
# annotated-types==0.7.0
# anyio==4.4.0
# asgiref==3.8.1
# attrs==24.2.0
# backoff==2.2.1
# beautifulsoup4==4.12.3
# certifi==2024.7.4
# chardet==5.2.0
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# coloredlogs==15.0.1
# dataclasses-json==0.6.7
# deepdiff==7.0.1
# Django==5.1
# emoji==2.12.1
# et-xmlfile==1.1.0
# faiss-cpu==1.8.0.post1
# filelock==3.15.4
# filetype==1.2.0
# flatbuffers==24.3.25
# frozenlist==1.4.1
# fsspec==2024.6.1
# greenlet==3.0.3
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.0
# httpx-sse==0.4.0
# huggingface-hub==0.24.5
# humanfriendly==10.0
# idna==3.7
# joblib==1.4.2
# jsonpatch==1.33
# jsonpath-python==1.0.6
# jsonpointer==3.0.0
# langchain==0.2.14
# langchain-community==0.2.12
# langchain-core==0.2.33
# langchain-mistralai==0.1.12
# langchain-text-splitters==0.2.2
# langdetect==1.0.9
# langsmith==0.1.99
# lxml==5.3.0
# Markdown==3.7
# marshmallow==3.21.3
# mistralai==0.4.2
# mpmath==1.3.0
# multidict==6.0.5
# mypy-extensions==1.0.0
# nest-asyncio==1.6.0
# networkx==3.3
# nltk==3.8.1
# numpy==1.26.4
# onnxruntime==1.19.0
# opencv-python==4.10.0.84
# openpyxl==3.1.5
# ordered-set==4.1.0
# orjson==3.10.7
# packaging==24.1
# pandas==2.2.2
# pillow==10.4.0
# protobuf==5.27.3
# psutil==6.0.0
# pyclipper==1.3.0.post5
# pydantic==2.8.2
# pydantic_core==2.20.1
# pypdf==4.3.1
# pyreadline3==3.4.1
# python-dateutil==2.9.0.post0
# python-docx==1.1.2
# python-dotenv==1.0.1
# python-iso639==2024.4.27
# python-magic==0.4.27
# pytz==2024.1
# PyYAML==6.0.2
# rapidfuzz==3.9.6
# # rapidocr-onnxruntime==1.3.24
# regex==2024.7.24
# requests==2.32.3
# requests-toolbelt==1.0.0
# shapely==2.0.5
# six==1.16.0
# sniffio==1.3.1
# soupsieve==2.6
# SQLAlchemy==2.0.32
# sqlparse==0.5.1
# sympy==1.13.2
# tabulate==0.9.0
# tenacity==8.5.0
# tokenizers==0.20.0
# tqdm==4.66.5
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# tzdata==2024.1
# unstructured==0.15.5
# unstructured-client==0.25.5
# urllib3==2.2.2
# wrapt==1.16.0
# xlrd==2.0.1
# yarl==1.9.4


